fixed bug where dequeueReuse cell was missing an 's' and giving me sibrt crash.  switched the location of where chart gets fed data to the AlertController handler and now app does exactly what its supposed to do.  only thing left is to incorporate CoreData to retrieve the proper saved timestamp for the tableView to reach mvp

+++ CalorieTrackerSPRINT/CalorieTrackerSPRINT/New Group/CalorieEntryRepresentation.swift
@@ -12,7 +12,7 @@ import Foundation
 
 class CalorieEntryRepresentation: Codable {
     
-    var calorie: Int16
+    var calorie: Double
     var timestamp: Date
     
 }
